% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mat.R
\name{combine.mat}
\alias{combine.mat}
\title{Combine function.}
\usage{
combine.mat(LL1, LL2)
}
\arguments{
\item{LL1}{List of SAVER results produced from first worker.}

\item{LL2}{List of SAVER results prodcued from another worker.}
}
\value{
A list of combined results
}
\description{
Combine function for \code{foreach} backend.
}
\details{
Genes are split across multiple workers with each worker outputing a list of
estimates, alpha, and beta values. This combine function combines the output
from the multiple workers into one list.
}
