% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds_eval.R
\name{.test_rds_eval}
\alias{.test_rds_eval}
\title{Test whether object can be loaded and its size}
\usage{
.test_rds_eval(test, filename, message_size)
}
\description{
Test whether object can be loaded and its size
}
