% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rds.r
\name{eval_rds}
\alias{eval_rds}
\title{Evaluate object of class "saver_uneval"}
\usage{
eval_rds(fn_or_call, p_dots, env_eval)
}
\value{
Output from evaluating object
}
\description{
Evaluate object of class "saver_uneval"
}
