% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu.R
\name{get.mu}
\alias{get.mu}
\title{Output prior predictions}
\usage{
get.mu(x, saver.obj)
}
\arguments{
\item{x}{Original count matrix.}

\item{saver.obj}{SAVER output.}
}
\value{
A matrix of prior mean predictions
}
\description{
Outputs prior predictions
}
\details{
This function outputs prior mean predictions \eqn{\mu} used in fitting the 
SAVER model.
}
\examples{
data("linnarsson")

saver1 <- saver(linnarsson, pred.genes = 1:5, pred.genes.only = TRUE)

mu <- get.mu(linnarsson[1:5, ], saver1)
mu[1:5, 1:5]

}
