% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linnarsson}
\alias{linnarsson}
\title{Mouse brain single-cell RNA-seq dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3529 rows and 200 columns.
}
\usage{
linnarsson
}
\description{
3,529 genes and 200 cells from a mouse brain scRNA-seq dataset.
}
\references{
Zeisel, A., Munoz-Manchado, A. B., Codeluppi, S.,
Lonnerberg, P., La Manno, G., Jureus, A., ... Linnarsson, S. (2015). Cell
types in the mouse cortex and hippocampus revealed by single-cell RNA-seq.
\emph{Science, 347}(6226), 1138-1142.
}
\keyword{datasets}
