% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linnarsson_saver}
\alias{linnarsson_saver}
\title{SAVER recovered mouse brain single-cell RNA-seq dataset}
\format{An object of class \code{saver} of length 3.}
\usage{
linnarsson_saver
}
\description{
Output of running `saver` on the `linnarsson` dataset.
}
\references{
Zeisel, A., Munoz-Manchado, A. B., Codeluppi, S.,
Lonnerberg, P., La Manno, G., Jureus, A., ... Linnarsson, S. (2015). Cell
types in the mouse cortex and hippocampus revealed by single-cell RNA-seq.
\emph{Science, 347}(6226), 1138-1142.
}
\keyword{datasets}
