% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rds_eval.r
\name{load_rds_eval}
\alias{load_rds_eval}
\title{Load an RDS object saved with save_rds_eval}
\usage{
load_rds_eval(filename, dir_save = NULL, eval = TRUE)
}
\description{
If object was not evaluated, it will be
before being returned.
}
