% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_variance.R
\name{calc.a}
\alias{calc.a}
\alias{calc.b}
\alias{calc.k}
\title{Optimizes variance}
\usage{
calc.a(y, mu, sf)

calc.b(y, mu, sf)

calc.k(y, mu, sf)
}
\arguments{
\item{y}{A vector of observed gene counts.}

\item{mu}{A vector of predictions from \code{\link{expr.predict}}.}

\item{sf}{Vector of normalized size factors.}
}
\value{
A vector with the optimized parameter and the negative
log-likelihood.
}
\description{
Finds the prior parameter that maximizes the marginal likelihood given
the prediction.
}
\details{
\code{calc.a} returns a prior alpha parameter assuming constant
coefficient of variation. \code{calc.b} returns a prior beta parameter
assuming constant Fano factor. \code{calc.k} returns a prior variance
parameter assuming constant variance.
}
