% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cols.R
\name{save_cols}
\alias{save_cols}
\title{Save Cols}
\usage{
save_cols(df, x)
}
\arguments{
\item{df}{Data frame}

\item{x}{Percentage of populated cols to keep, as a decimal.}
}
\value{
Data frame
}
\description{
Save Cols
}
\examples{
df<-data.frame(
col1=c('cat','dog','bird','fish'),
col2=c(2,3,NA,NA),
col3=c('a','b','c','d'))

result_df<-save_cols(df,0.33)

}
