% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds_eval.R
\name{save_rds_eval}
\alias{save_rds_eval}
\title{Save objects as rds without enclosing extra objects}
\usage{
save_rds_eval(
  fn_or_call = NULL,
  filename,
  dir_save = NULL,
  return_obj = FALSE,
  eval_fun = TRUE,
  test = FALSE,
  message_size = 1,
  ...
)
}
\arguments{
\item{fn_or_call}{function or call.
Function whose output is the `ggplot` object.}

\item{filename}{character.
File name of output object.
Note that if it does not end in \code{".rds"}, \code{".rds"} is appended.}

\item{dir_save}{character.
Directory to save to.
Default is \code{NULL}, in which case the save directory is
\code{here::here()}.}

\item{return_obj}{logical.
If \code{TRUE}, then the output of \code{fn_or_call} is returned.
Default is \code{FALSE}.}

\item{eval_fun}{logical.
If \code{FALSE}, then \code{fn_or_call}
is not evaluated even if it is not \code{NULL}.
Default is \code{TRUE}.}

\item{test}{logical.
If \code{FALSE} and \code{eval_fun == FALSE}, then
\code{load_rds_eval} is run to see if the
saved object can be used in evaluation without
returning an error. In particular, this is used
to check that everything that should be saved
via \code{...} is saved. Default is \code{FALSE}.}

\item{message_size}{numeric. If \code{test == TRUE},
then a message is printed if the size exceeds \code{warning} in MB.
Default is 1.
Set to 0 to always message the size if \code{test == TRUE}.}

\item{...}{name-value pairs (\code{alist}).
These are passed to code{fn_or_call} as parameter-argument pairs.}
}
\value{
If \code{return_obj == FALSE}, then \code{invisible(TRUE)}.
If \code{return_obj == TRUE}, then output type depends on \code{eval_fun}.
If \code{eval_fun == TRUE}, then the output from
evaluating \code{fn_or_call}.
If \code{eval_fun == FALSE}, then
\code{list(fn_or_call = fn_or_call, p_dots = p_dots)}.
}
\description{
Many objects, such as \code{ggplot2::gplot} objects and formulae,
enclose the \href{https://adv-r.hadley.nz/environments.html#execution-environments}{calling environment}.
If such objects are saved using \code{saveRDS},
then the enclosed environments are also saved.
Therefore this may unintentionally save large objects.
This function avoids this by evaluating the function in
the global environment (or a direct child), and objects in
global environments are not saved when enclosed.
The outputted object can then be saved without extra unwanted items.
}
\details{
For example, \code{ggplot2::gplot} encloses the
calling environment in two places: the \code{mapping}
and \code{plot_env} list elements.
}
\examples{
plot_fn <- function() {
 ggplot(data.frame(x = 1, y = 1), aes(x,y)) + geom_point()
}
save_rds(plot_fn,
         "example.rds",
         return_obj = TRUE)
p <- readRDS("example.rds")
p
pryr::object_size(p)

}
