% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rows.R
\name{save_rows}
\alias{save_rows}
\title{Save Rows}
\usage{
save_rows(df, x)
}
\arguments{
\item{df}{Data frame object with rows and columns}

\item{x}{Percentage of populated rows to keep, as a decimal.}
}
\value{
Data frame
}
\description{
Save Rows
}
\examples{
df<-data.frame(
col1=c('cat','dog','bird','fish'),
col2=c(2,3,NA,NA)
col3=c('a','b','c','d'))

result_df<-save_rows(df,0.33)
}
