% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVER-package.R
\docType{package}
\name{SAVER-package}
\alias{SAVER-package}
\title{SAVER: Single-cell Analysis Via Expression Recovery}
\source{
\url{https://github.com/mohuangx/SAVER}
}
\description{
The SAVER package implements SAVER, a gene expression recovery method for
single-cell RNA sequencing (scRNA-seq). Borrowing information across all
genes and cells, SAVER provides estimates for true expression levels as well
as posterior distributions to account for estimation uncertainty.
}
\author{
Mo Huang, \url{mohuang@wharton.upenn.edu}
}
