% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_predict.R
\name{expr.predict}
\alias{expr.predict}
\title{Calculates SAVER prediction.}
\usage{
expr.predict(x, y, dfmax = 300, nfolds = 5, seed = NULL, cutoff = 0)
}
\arguments{
\item{x}{A log-normalized expression count matrix of genes to be used in the
prediction.}

\item{y}{A normalized expression count vector of the gene to be predicted.}

\item{dfmax}{The number of genes to be included in the prediction. Default
is 300.}

\item{nfolds}{Number of folds to use in the cross-validation. Default is 5.}

\item{seed}{Sets the seed for reproducible results.}

\item{cutoff}{If the model with the lowest mean cross-validation error has
Poisson deviance greater than the mean cross-validation error of the null
model subtracted by the estimated standard error of the mean
cross-validation error of the null model times \code{cutoff}, then the null
model is selected. Default is 0.}
}
\value{
A vector of predicted gene expression.
}
\description{
Uses \code{cv.glmnet} from the \code{glmnet} package to return the SAVER
prediction.
}
\details{
The SAVER method starts with predicting the prior mean for each cell for a
specific gene. The prediction is performed using the observed normalized
gene count as the response and the normalized gene counts of other genes as
predictors. \code{cv.glmnet} from the \code{glmnet} package is used to fit
the LASSO Poisson regression. The model with the lowest cross-validation
error is chosen and the fitted response values are returned and used as the
SAVER prediction.
}
