% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu.R
\name{get.mu}
\alias{get.mu}
\title{Output prior predictions}
\usage{
get.mu(x, saver.obj)
}
\arguments{
\item{x}{Original count matrix.}

\item{saver.obj}{SAVER output.}
}
\value{
A matrix of prior mean predictions
}
\description{
Outputs prior predictions
}
\details{
This function outputs prior mean predictions \eqn{\mu} used in fitting the 
SAVER model.
}
\examples{
data("linnarsson")
data("linnarsson_saver")

mu <- get.mu(linnarsson, linnarsson_saver)

}
