% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVER-package.R
\docType{package}
\name{SAVER-package}
\alias{SAVER-package}
\title{SAVER: Single-cell Analysis Via Expression Recovery}
\source{
\url{https://github.com/mohuangx/SAVER}
}
\description{
The SAVER package implements SAVER, a gene expression recovery method for
single-cell RNA sequencing (scRNA-seq). Borrowing information across all
genes and cells, SAVER provides estimates for true expression levels as well
as posterior distributions to account for estimation uncertainty. See
\href{https://doi.org/10.1038/s41592-018-0033-z}{Huang et al (2018)} for
more details.
}
\author{
Mo Huang (Maintainer), \email{mohuangx@gmail.com}

Nancy Zhang, \email{nzh@wharton.upenn.edu}

Mingyao Li, \email{mingyao@pennmedicine.upenn.edu}
}
